separate array and struct into different packages and move both as type's sub packages
