+++ chameleon-core/src/main/java/org/hibernate/omm/dialect/MongoDialect.java
@@ -32,10 +32,10 @@ import org.hibernate.mapping.Constraint;
 import org.hibernate.mapping.Index;
 import org.hibernate.mapping.Selectable;
 import org.hibernate.mapping.Table;
-import org.hibernate.omm.array.function.MongoArrayContainsFunction;
-import org.hibernate.omm.array.function.MongoArrayIncludesFunction;
-import org.hibernate.omm.array.type.MongoSQLArrayJdbcTypeConstructor;
-import org.hibernate.omm.array.type.MongoSQLStructCastingJdbcType;
+import org.hibernate.omm.type.array.function.MongoArrayContainsFunction;
+import org.hibernate.omm.type.array.function.MongoArrayIncludesFunction;
+import org.hibernate.omm.type.array.MongoSQLArrayJdbcTypeConstructor;
+import org.hibernate.omm.type.struct.MongoSQLStructCastingJdbcType;
 import org.hibernate.omm.ast.MQLAstTranslatorFactory;
 import org.hibernate.omm.dialect.exporter.MongoIndexCommandUtil;
 import org.hibernate.omm.type.ObjectIdJavaType;
