+++ chameleon-core/src/main/java/org/hibernate/omm/jdbc/MongoConnection.java
@@ -20,7 +20,7 @@ import com.mongodb.client.ClientSession;
 import com.mongodb.client.MongoDatabase;
 import com.mongodb.lang.Nullable;
 import org.bson.Document;
-import org.hibernate.omm.array.MongoArray;
+import org.hibernate.omm.type.array.MongoArray;
 import org.hibernate.omm.exception.NotYetImplementedException;
 import org.hibernate.omm.jdbc.adapter.ConnectionAdapter;
 import org.hibernate.omm.jdbc.exception.CommandRunFailSQLException;
