+++ chameleon-core/src/main/java/org/hibernate/omm/type/array/MongoSQLArrayJdbcType.java
@@ -16,7 +16,7 @@
  *
  */
 
-package org.hibernate.omm.array.type;
+package org.hibernate.omm.type.array;
 
 import org.hibernate.HibernateException;
 import org.hibernate.engine.spi.SharedSessionContractImplementor;
